move unix/helper to utils dir
